Moved cld2 unit tests to separate file and instead use a more straightforward LanguageDetector DOM sample based test approach
